+++ NEWS.md
@@ -1,5 +1,9 @@
 # dplyr (development version)
 
+## Minor improvements and bug fixes
+
+* cummean() no longer has off-by-one indexing problem (#5287).
+
 # dplyr 1.0.0
 
 ## Breaking changes