+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * cummean() no longer has off-by-one indexing problem (#5287).
 
+* The call stack is preserved on error. This makes it possible to `recover()`
+  into problematic code called from dplyr verbs (#5308).
+
+
 # dplyr 1.0.0
 
 ## Breaking changes