+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## Minor improvements and bug fixes
 
-* cummean() no longer has off-by-one indexing problem (#5287).
+* `coalesce()` now supports data frames correctly (#5326).
+
+* `cummean()` no longer has off-by-one indexing problem (#5287).
 
 # dplyr 1.0.0