+++ NEWS.md
@@ -1,4 +1,4 @@
-# dplyr (development version)
+# dplyr 1.0.1
 
 * New function `cur_data_all()` similar to `cur_data()` but includes the grouping variables (#5342).